Add files via upload
This script is used for quick backups of MySQL databases during the development of projects for Primal eCommerce. Python version is 3.12 Don't forget to create a python virtual environment at the directory running the script. An additional prerequisite is to create an SSH connection to all hosts running MySQL or MariaDB Databases before running the script.